requirements.txt.txt
Flask
pycryptodome
gunicorn
